create method verify remote att
