--- crates/remote-attestation-verifier/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "remote-attestation-verifier"
-version = "0.1.0"
-authors = ["asa93"]
-edition = "2018"
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde_cbor = { version = "0.11.2", default-features = true, features = [
-    "alloc",
-] }
-serde = "1.0"
-serde_with = { version = "1.9.4", default-features = false, features = [] }
-base64 = "0.22.1"
-x509-cert = { version = "0.2.5", default-features = false }
-rsa = "0.8.0"
-pem = "3.0.1"
-p384 = { version = "0.13.0", default-features = false, features = [
-    "arithmetic",
-    "ecdh",
-    "ecdsa",
-    "test-vectors",
-    "expose-field",
-    "pkcs8",
-    "sha384",
-] }
-cbor = "0.3"
-hex = "0.4.3"
-rand_core = { version = "0.6.4", features = ["getrandom"] }
-
-#ciborium = "0.2.2"
-
-#reqwest = { version = "0.10.10", features = ["blocking"] }
-#serde_bytes = "0.11.15"
-#rustls = { version = "0.20.4", features = ["dangerous_configuration"] }
-#webpki = "0.22.0"
-#byteorder = "1.3.4"
